Design and evaluation of a closed-loop deep brain stimulation system for movement and mental disorders
In contrast to the standard form of continuous deep brain stimulation
that is used clinically for the treatment of motor and psychiatric
disorders, in a closed-loop system, the characteristics of the
excitation are formed in real time depending on the clinical condition
of the patient, as captured by predetermined neurophysiological
biomarkers. The validation of the effectiveness of this system would be
equivalent to both the minimization of energy requirements associated
with the application of deep brain stimulation, and the gradual
displacement of the very lengthy process of postoperative clinical
programming.
Purpose of this thesis is the design and optimization of a closed-loop
deep brain stimulation system by means of a stochastic dynamical model
and adaptation of this system to intraoperative microelectrode recording
data, to treat Parkinson's disease and obsessive-compulsive disorder.